pyproject.toml
[build-system]
requires = ["setuptools","wheel","numpy"]
build-backend = "setuptools.build_meta"

[project]
name = "snpy"
version = "2.7.2"
authors = [
    {name="Chris Burns", email="cburns@carnegiescience.edu"},
]
description = "SNooPy:  Supernova light-curve analysis tool"
readme = "README.rst"
requires-python = ">=3.8"
keywords = ["astronomy","supernova","science"]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Environment :: Console",
    "Framework :: IPython",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 2",
    "Programming Language :: Python :: 2.6",
    "Programming Language :: Python :: 2.7",
    "Topic :: Scientific/Engineering :: Astronomy"
]
dependencies = [
    "oldest-supported-numpy",
    "scipy",
    "astropy",
    "ipython",
    "matplotlib",
    "pandas",
    "pymysql",
    "scikit-learn",
    "h5py",
    "onnxruntime",
    "emcee"]


[tool.setuptools]
packages = ["snpy","snpy.CSPtemp","snpy.SwiftTemp","snpy.dm15temp",
            "snpy.filters","snpy.spline2","snpy.sqlmod","snpy.utils"]

[tool.setuptools.package-data]
"snpy.CSPtemp" = ["fits/*.fits"]
"snpy.dm15temp" = ["*.pickle"]
"snpy.utils" = ["*.pickle"]
"snpy.filters" = ["filters/*/*/*","standards/*/*"]
